DAO Tests Setup & Cleanup Implemented

+++ esp-src/esp-be/src/main/java/edu/esp/database/doas/AdminDAO.java
@@ -2,15 +2,9 @@ package edu.esp.database.doas;
 
 import edu.esp.system_entities.system_users.Admin;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
-
-import edu.esp.system_entities.system_users.Admin;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.jdbc.core.RowMapper;
-
 import java.util.List;
 
 public class AdminDAO {
@@ -20,7 +14,7 @@ public class AdminDAO {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Admin readAdminById(int id) {
+    public Admin readAdminById(byte id) {
         try {
             String sql = """
                 SELECT *
@@ -67,7 +61,7 @@ public class AdminDAO {
 
     }
 
-    public boolean deleteAdminById(int id) {
+    public boolean deleteAdminById(byte id) {
         try{
             String sql = """
                     DELETE FROM sys_admin
